# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CompareIn": {
        "additionalProperties": false,
        "properties": {
          "query": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "CompareIn",
        "type": "object"
      },
      "CompareOut": {
        "properties": {
          "agreement": {
            "description": "True when every strategy picked the same agent.",
            "title": "Agreement",
            "type": "boolean"
          },
          "decisions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DecisionOut"
            },
            "title": "Decisions",
            "type": "object"
          },
          "embedding_latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Latency Ms"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query",
          "embedding_latency_ms",
          "decisions",
          "agreement"
        ],
        "title": "CompareOut",
        "type": "object"
      },
      "DecisionOut": {
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateOut"
            },
            "title": "Candidates",
            "type": "array"
          },
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "embedding_latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Latency Ms"
          },
          "fallback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "reasoning": {
            "title": "Reasoning",
            "type": "string"
          },
          "strategy": {
            "title": "Strategy",
            "type": "string"
          }
        },
        "required": [
          "strategy",
          "agent",
          "fallback",
          "reasoning",
          "latency_ms",
          "cost_usd",
          "candidates",
          "metadata"
        ],
        "title": "DecisionOut",
        "type": "object"
      },
...
      "FeedbackIn": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "success": {
            "description": "Did the answer solve the task?",
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "FeedbackIn",
        "type": "object"
      },
...
      "ProblemOut": {
        "description": "RFC 9457 problem details, returned for every error.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Validation errors (422).",
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "instance",
          "request_id"
        ],
        "title": "ProblemOut",
        "type": "object"
      },
      "QueryCreate": {
        "additionalProperties": false,
        "examples": [
          {
            "mode": "sync",
            "query": "What is 17.5% of 640?",
            "strategy": "adaptive"
          }
        ],
        "properties": {
          "execute": {
            "default": true,
            "description": "False = route only, don't run the agent.",
            "title": "Execute",
            "type": "boolean"
          },
          "mode": {
            "default": "sync",
            "description": "async returns 202 immediately; poll GET /v1/queries/{id}.",
            "enum": [
              "sync",
              "async"
            ],
            "title": "Mode",
            "type": "string"
          },
          "query": {
            "description": "The user's request.",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "strategy": {
            "anyOf": [
              {
                "enum": [
                  "round_robin",
                  "embedding",
                  "llm",
                  "adaptive"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Routing strategy; defaults to the server's default.",
            "title": "Strategy"
          },
          "use_cache": {
            "default": true,
            "description": "Allow serving an identical cached answer.",
            "title": "Use Cache",
            "type": "boolean"
          }
        },
        "required": [
          "query"
        ],
        "title": "QueryCreate",
        "type": "object"
      },
      "QueryList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuerySummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `before` to fetch the next page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "QueryList",
        "type": "object"
      },
      "QueryOut": {
        "properties": {
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Output of the final successful execution, if any.",
            "title": "Answer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionOut"
          },
          "executions": {
            "items": {
              "$ref": "#/components/schemas/ExecutionOut"
            },
            "title": "Executions",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Links",
            "type": "object"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "status": {
            "enum": [
              "routed",
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "total_cost_usd": {
            "description": "Router + all agent executions (estimated).",
            "title": "Total Cost Usd",
            "type": "number"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "id",
          "query",
          "status",
          "created_at",
          "completed_at",
          "decision",
          "executions",
          "answer",
          "total_cost_usd",
          "trace_id",
          "links"
        ],
        "title": "QueryOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nAdaptiveRoute routes each query to one of five specialised LLM agents\n(code, math, sql, writer, knowledge) using one of four strategies:\n**round_robin**, **embedding**, **llm** and **adaptive**.\n\nAuthenticate with `Authorization: Bearer <api key>` (or `X-API-Key`). Errors use\n`application/problem+json`. POST /v1/queries accepts an `Idempotency-Key` header.\n",
    "title": "AdaptiveRoute API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/queries": {
      "get": {
        "description": "Newest first, keyset-paginated. Admins see every query; users see their own.",
        "operationId": "list_queries_v1_queries_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from `next_cursor`",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from `next_cursor`",
              "title": "Before"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Validation error (or Idempotency-Key reused with a different body)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Rate limit exceeded (see Retry-After)"
          }
        },
        "summary": "List recent queries",
        "tags": [
          "queries"
        ]
      },
      "post": {
        "description": "Routes the query with the chosen strategy, stores the decision and, unless\n`execute=false`, runs the selected agent. With `mode=async` the agent runs on the\nCelery worker and the response is `202` with `status=queued`.\n\nSend an `Idempotency-Key` header to make retries safe: a repeated request with the\nsame key returns the stored response instead of executing again.",
        "operationId": "create_query_v1_queries_post",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryOut"
                }
              }
            },
            "description": "Accepted for async execution"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Validation error (or Idempotency-Key reused with a different body)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Rate limit exceeded (see Retry-After)"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Timed out (includes query_id)"
          }
        },
        "summary": "Route a query to an agent (and optionally execute it)",
        "tags": [
          "queries"
        ]
      }
    },
    "/v1/queries/{query_id}": {
      "get": {
        "operationId": "get_query_v1_queries__query_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "query_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Query Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Validation error (or Idempotency-Key reused with a different body)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Rate limit exceeded (see Retry-After)"
          }
        },
        "summary": "Get a query and its result",
        "tags": [
          "queries"
        ]
      }
    },
    "/v1/queries/{query_id}/feedback": {
      "post": {
        "description": "Labels the query's final execution as a success or failure. These labels are the\ntraining signal for the adaptive router's historical-success term.",
        "operationId": "give_feedback_v1_queries__query_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "query_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Query Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Validation error (or Idempotency-Key reused with a different body)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Rate limit exceeded (see Retry-After)"
          }
        },
        "summary": "Record whether the answer solved the task",
        "tags": [
          "queries"
        ]
      }
    },
    "/v1/route/compare": {
      "post": {
        "description": "Shows which agent each strategy would choose and why. Nothing is executed or\nstored; the LLM router's call is billed as usual.",
        "operationId": "compare_strategies_v1_route_compare_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Validation error (or Idempotency-Key reused with a different body)"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Rate limit exceeded (see Retry-After)"
          }
        },
        "summary": "Run every routing strategy on a query (no execution)",
        "tags": [
          "queries"
        ]
      }
    },
...
  },
...
}